requirements.txt
pycryptodome==3.12.0
cryptography==2.7
pyOpenSSL==19.0.0
dgl==2.1.0
sklearn-pandas==2.2.0
torch==2.2.1